# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Finna FinOps API
  version: 0.2.0
  description: 'The Finna FinOps API is the backend for the Finna cloud cost management platform. It handles multi-cloud cost
    ingestion (Azure Cost Management, GCP Billing, LLM usage), resource wastage detection, alert management, and per-project
    FinOps dashboards. Extracted cost records are normalised into a unified schema and made available through a rich query
    layer with time-window shortcuts (`mtd`, `7d`, `30d`, `90d`), provider filters, and CSV export.


    Authentication supports three flows: username/password JWT (primary), GitHub OAuth (developer login), and pluggable OIDC
    providers with PKCE and auto-provisioning. All resource endpoints require a valid JWT Bearer token. Admin-only endpoints
    (auth provider CRUD) require an `is_admin` claim in the token.


    This repository contains the API-only backend. The React/TypeScript UI lives in a separate repo and communicates exclusively
    through this API.'
  contact:
    name: Finna Team
    email: andrea.carmisciano@abstract.it
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /api/v1/config:
    get:
      tags:
      - Config
      summary: List cloud credential configs
      operationId: list_configs_api_v1_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CloudConfigResponse'
                type: array
                title: Response List Configs Api V1 Config Get
      security:
      - OAuth2PasswordBearer: []
      description: Returns all cloud provider credential configurations. Sensitive fields (`client_secret`, `key_file_content`)
        are redacted. Configs are decrypted from storage before masking.
    post:
      tags:
      - Config
      summary: Create cloud credential config
      operationId: create_config_api_v1_config_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudConfigCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConfigResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      description: Stores a new cloud credential entry (provider, credential type, config blob) encrypted at rest. Returns
        the created record with secrets masked.
  /api/v1/config/projects:
    get:
      tags:
      - Config
      summary: List projects with MTD cost
      description: Returns all `fin_projects` entries joined with cost records for the requested time window. The `mtd` field
        is the sum of costs in the period, defaulting to month-to-date.
      operationId: list_projects_api_v1_config_projects_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: window
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Time window: mtd, 7d, 30d, 90d'
          title: Window
        description: 'Time window: mtd, 7d, 30d, 90d'
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Start date in ISO format
          title: Start Date
        description: Start date in ISO format
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: End date in ISO format
          title: End Date
        description: End date in ISO format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Projects Api V1 Config Projects Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Config
      summary: Create project
      description: Creates a new FinOps project entry. Slug is auto-derived from the name if not provided. Used to track cost
        centre attribution.
      operationId: create_project_api_v1_config_projects_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Data
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Project Api V1 Config Projects Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/config/projects/{slug}:
    get:
      tags:
      - Config
      summary: Get project by slug
      description: Fetches a single project by its slug, joined with cost records for the requested window.
      operationId: get_project_api_v1_config_projects__slug__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      - name: window
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Time window: mtd, 7d, 30d, 90d'
          title: Window
        description: 'Time window: mtd, 7d, 30d, 90d'
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Start date in ISO format
          title: Start Date
        description: Start date in ISO format
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: End date in ISO format
          title: End Date
        description: End date in ISO format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Project Api V1 Config Projects  Slug  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Config
      summary: Delete project
      description: Removes a project entry by slug. Cost records referencing this project's slug are not deleted.
      operationId: delete_project_api_v1_config_projects__slug__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/config/{config_id}/test:
    post:
      tags:
      - Config
      summary: Test cloud credentials
      description: Verifies that the stored credentials can authenticate and access the cloud API. For Azure, performs a Cost
        Management query against the configured subscription scope. For GCP, refreshes ADC credentials.
      operationId: test_config_api_v1_config__config_id__test_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: config_id
        in: path
        required: true
        schema:
          type: string
          title: Config Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Test Config Api V1 Config  Config Id  Test Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/configs:
    get:
      tags:
      - Config
      summary: CLI-compatible config list
      description: Alias of `GET /config` returning a paginated wrapper object (`data`, `total`, `page`). Intended for the
        Finna CLI tool.
      operationId: cli_config_list_api_v1_configs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Cli Config List Api V1 Configs Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - Config
      summary: CLI-compatible config create
      description: Simplified config create for the CLI. Accepts `provider`, `service_category`, `region`, and `cloud_config`
        fields.
      operationId: cli_configs_create_api_v1_configs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_ConfigCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Cli Configs Create Api V1 Configs Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    CloudConfigResponse:
      properties:
        id:
          type: string
          title: Id
        provider:
          type: string
          title: Provider
        name:
          type: string
          title: Name
        credential_type:
          type: string
          title: Credential Type
        config:
          additionalProperties: true
          type: object
          title: Config
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        last_test:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Test
        last_test_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Test At
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
        subscription_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Subscription Id
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
        err:
          anyOf:
          - type: string
          - type: 'null'
          title: Err
      type: object
      required:
      - id
      - provider
      - name
      - credential_type
      - config
      - created_at
      - updated_at
      title: CloudConfigResponse
      description: Response schema for cloud configuration.
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/token
servers:
- url: http://localhost:8000
  description: Local development server
...